Added async dumping mode, added CLRMD with patch that supports .NET 4.6

+++ DumpWriter/Native.cs
@@ -252,7 +252,7 @@ namespace DumpWriter
         public IntPtr Handle;
         public ulong Offset;
         public IntPtr Buffer;
-        public ulong BufferBytes;
+        public uint BufferBytes;
     }
 
     [StructLayout(LayoutKind.Sequential, Pack = 4)]
